requirements.txt
Click==7.0
decorator==4.4.0
Flask==1.1.1
Flask-WTF==0.14.2
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dotenv==0.10.3
six==1.12.0
validators==0.14.0
Werkzeug==0.15.6
WTForms==2.2.1
WTForms-Components==0.10.4
